Designer: Make the NeuroRT Engine Control Panel great again
This modifies the behaviour of the NeuroRT Engine control panel in several ways:
- The code now searches for the presence of the mensia-engine library in PATH and loads it dynamically
- Symbols are loaded using DynamicModule, the mensia headers are no longer necessary and the API is no longer fragile

The module is always built, both for open source and mensia version, this can be changed later. If the library is not found
then the failure is silent as to not pollute the console with useless messages.

Bump version to 0.0.5

Tracker Reference: [CORE-837](https://jira.mensiatech.com/browse/CORE-837)