[Designer] Cache deprecated and present box status in the proxy
There is little chance that a box will suddenly become deprecated when
editing a scenario. It can theoretically disappear if one deletes the
metabox scenario under the Designer's feet but this is an edge case and
not worth the overhead of always checking for the file and algorithm
presence.

This does not eliminate the massive overhead the BoxProxy creates as it
is still recreated on every redraw of the scenario.